requirements.txt
dify_plugin>=0.2.0,<0.3.0
weaviate-client>=4.6.0
numpy>=1.24.0
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.7.0
